Move translate_table_columns to expr.rs
